Update folder structure, import order
