+++ src/app/components/selected-clue/selected-clue.component.html
