+++ src/app/pages/answer-drafting/answer-drafting.component.scss
