+++ src/app/pages/clue-editing/clue-editing.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from "@angular/core";
 import { FormArray, FormControl, FormGroup } from "@angular/forms";
-import { Clue, ClueType, PuzzleService } from "../services/puzzle.service";
+
+import { PuzzleService } from "src/app/services/puzzle.service";
+import { Clue, ClueType } from "src/app/models/clue.model";
 
 @Component({
   selector: "app-clue-editing",
