+++ src/app/pages/load-puzzle/load-puzzle.component.scss
