+++ src/app/pages/puzzle-editing/puzzle-editing.component.scss
