+++ src/app/pages/puzzle-review/puzzle-review.component.scss
