+++ src/app/pages/puzzle-stats/puzzle-stats.component.scss
