+++ src/app/pages/sign-in/sign-in.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from "@angular/core";
-import { AuthService } from "../services/auth.service";
 import { FormControl, FormGroup, Validators } from "@angular/forms";
-import { FirebaseError } from "@angular/fire/app";
 import { Router } from "@angular/router";
+
+import { AuthService } from "src/app/services/auth.service";
 
 @Component({
   selector: "app-header-nav",
