+++ src/app/services/puzzle.service.ts
@@ -1,115 +1,11 @@
 import { Injectable } from "@angular/core";
 import { BehaviorSubject, Observable } from "rxjs";
 import { catchError, switchMap } from "rxjs/operators";
+
 import { SaveService } from "./save.service";
 import { LoadService } from "./load.service";
-
-export interface PuzzleDoc {
-  id: string;
-  name: string;
-  createdBy: string;
-  width: number;
-  height: number;
-  locked: boolean;
-  answers: Array<string>;
-  spacers: Array<number>;
-  circles: Array<number>;
-  shades: Array<number>;
-  "across-clues": Array<string>;
-  "down-clues": Array<string>;
-}
-
-export class Puzzle {
-  id: string;
-  name: string;
-  createdBy: string;
-  width: number;
-  height: number;
-  locked: boolean;
-  grid: Array<Square>;
-  acrossClues: Array<Clue>;
-  downClues: Array<Clue>;
-
-  constructor(
-    id: string = "",
-    name: string = "",
-    createdBy: string = "",
-    width: number = 0,
-    height: number = 0,
-    locked: boolean = false,
-    grid: Array<Square> = [],
-    acrossClues: Array<Clue> = [],
-    downClues: Array<Clue> = []
-  ) {
-    this.id = id;
-    this.name = name;
-    this.createdBy = createdBy;
-    this.locked = locked;
-    this.width = width;
-    this.height = height;
-    this.grid = grid;
-    this.acrossClues = acrossClues;
-    this.downClues = downClues;
-  }
-}
-
-export class Clue {
-  num: number;
-  text: string;
-  answer: string;
-  squares: Array<number>;
-
-  constructor(num: number = -1, text: string = "", answer: string = "", squares: Array<number> = []) {
-    this.num = num;
-    this.text = text;
-    this.answer = answer;
-    this.squares = squares;
-  }
-}
-
-export enum ClueType {
-  Across,
-  Down,
-}
-
-export class Square {
-  index: number;
-  value: string;
-  type: SquareType;
-  overlay: OverlayType;
-  boxNum: number;
-  downClueNum: number;
-  acrossClueNum: number;
-
-  constructor(
-    index: number = -1,
-    value: string = " ",
-    boxNum: number = -1,
-    acrossClueNum: number = -1,
-    downClueNum: number = -1,
-    type: SquareType = SquareType.Letter,
-    overlay: OverlayType = OverlayType.None
-  ) {
-    this.index = index;
-    this.value = value;
-    this.boxNum = boxNum;
-    this.acrossClueNum = acrossClueNum;
-    this.downClueNum = downClueNum;
-    this.type = type;
-    this.overlay = overlay;
-  }
-}
-
-export enum SquareType {
-  Letter,
-  Spacer,
-}
-
-export enum OverlayType {
-  None,
-  Circle,
-  Shade,
-}
+import { Clue, ClueType } from "../models/clue.model";
+import { PuzzleDoc, Puzzle, Square, SquareType, OverlayType } from "../models/puzzle.model";
 
 @Injectable({
   providedIn: "root",
